requirements.txt
discord
openai
Pillow
python-dotenv